Node parsing: Completes parsing of nodes from a file.
NEXT: Perform some sanity checks on the nodes, execute what can be
executed, etc.

NOTE: The specification isn't updated as yet - I want special
`entry-point` nodes. Currently only the `cron` node is an entry-point.
Job nodes should not be entry-points purely because if they are then it
is not feasiable to check the execution tree for errors.

With the planned entry-point nodes of cron, github/webhook,
gitlab/webhook, c/line webhook and a gitea webhook, should be enough to
get started.