Single-shot test: Adds single-shot tests.
Tests are added, updated expected outputs. For each failing test,
add in the functionality, then check the output, then copy it (if
correct) to the tests/expected/ directory.

Tests all run with --lint, to allow completion.

EMIT and HANDLES implemented
EXEC also implemented, not well tested with shell pipelined commands

NEXT: Implement a check for any node with an EMITS to determine if a
HANDLES node exists for that signal. This must be done in kbtree.c (see
TODO item in the file in tree_eval()).

NEXTNEXT: Maybe look into allowing multiple EXEC to be specified via the
array syntax. The commands will be executed in order, and only proceed
to the next command when the previous command returns success.