Adds test and code to ensure every EMIT is handled.
Also updates exec output, standardises on '^::' as marker for kubeka
output, separated with colons to facilitate sed/cut uses.

Updates output to match the '^::' standard, updates tests to match.

NEXT: Test appears to be failing, even though the correct output is
generated. This is probably due to not having a tests/expected... file
for the unhandled-signal test.

NEXTNEXT: Fixup the errorring single-* tests. Maybe they are checking
for success and not failure, or perhaps the tests/expected/* files don't
exist yet.

NEXTNEXNEXT: Attempt to use this as a release 1.0.