Adds periodic support via threads.
NOTE: There is a bug, the counter variable is not updated. The
way to fix this would be to reinstantiate, then reevaluate the
node as specified by the node id.

This means that all the thread functions need to take the dedeup_nodes
collection and a node id, and not the root node.

This is a tedious change, will mark it as a bug in the GH issues page.